frontend: helpers: Move loadTableSettings to ResourceTable
It's not used anywhere else.

Also move storeTableSettings there.
