+++ frontend/src/helpers/index.ts
@@ -55,31 +55,6 @@ function getProductName(): string | undefined {
   return import.meta.env.REACT_APP_HEADLAMP_PRODUCT_NAME;
 }
 
-function storeTableSettings(tableId: string, columns: { id?: string; show: boolean }[]) {
-  if (!tableId) {
-    console.debug('storeTableSettings: tableId is empty!', new Error().stack);
-    return;
-  }
-
-  const columnsWithIds = columns.map((c, i) => ({ id: i.toString(), ...c }));
-  // Delete the entry if there are no settings to store.
-  if (columnsWithIds.length === 0) {
-    localStorage.removeItem(`table_settings.${tableId}`);
-    return;
-  }
-  localStorage.setItem(`table_settings.${tableId}`, JSON.stringify(columnsWithIds));
-}
-
-function loadTableSettings(tableId: string): { id: string; show: boolean }[] {
-  if (!tableId) {
-    console.debug('loadTableSettings: tableId is empty!', new Error().stack);
-    return [];
-  }
-
-  const settings = JSON.parse(localStorage.getItem(`table_settings.${tableId}`) || '[]');
-  return settings;
-}
-
 const exportFunctions = {
   getBaseUrl,
   isDevMode,
@@ -95,8 +70,6 @@ const exportFunctions = {
   storeClusterSettings,
   loadClusterSettings,
   getHeadlampAPIHeaders,
-  storeTableSettings,
-  loadTableSettings,
 };
 
 export default exportFunctions;
